requirements.txt
tensorflow
svgwrite
svglib
reportlab
Pillow
IPython
pdf2image
cairosvg
google-colab
